+++ gorgon/gui/skeletonization/binary.py
