+++ gorgon/gui/skeletonization/grayscale.py
